example/functions/img-resize-fn-py/requirements.txt
# requests
fastapi~=0.110.0
aiohttp~=3.9.3
uvicorn~=0.28.1
gunicorn
oaas-sdk-py==0.2.3
pillow~=10.2.0

aiofiles~=23.2.1
oaas_sdk_py~=0.2.3
